+++ src/arcrest/packages/ntlm3/compat.py
@@ -2,6 +2,6 @@
 
 def _long(value):
     try:
-        return long(value)
+        return int(value)
     except NameError:  # we're Python 3, we don't have longs
         return int(value)
