+++ src/arcrest/webmap/symbols.py
@@ -1,5 +1,5 @@
-from __future__ import absolute_import
-from __future__ import print_function
+
+
 from .._abstract.abstract import BaseSymbol
 import os
 import json
@@ -206,7 +206,7 @@ class SimpleMarkerSymbol(BaseSymbol):
     @xoffset.setter
     def xoffset(self, value):
         """ sets the xoffset """
-        if isinstance(value (int, float, long)):
+        if isinstance(value (int, float, int)):
             self._xoffset = value
     #----------------------------------------------------------------------
     @property
@@ -217,7 +217,7 @@ class SimpleMarkerSymbol(BaseSymbol):
     @yoffset.setter
     def yoffset(self, value):
         """ sets the y offset """
-        if isinstance(value (int, float, long)):
+        if isinstance(value (int, float, int)):
             self._yoffset = value
     #----------------------------------------------------------------------
     @property
